Reduce excessive (1mil/s) file watcher frequency
It looks like the FileWatcher is intended to sleep for 1 second (1000ms)
before re-checking watched files, but it actually sleeps for 1000ns - so
it checks watched files up to 1 million times every second! This causes
a fair amount of excessive CPU load, even while the process is idle.

Renamed SLEEP_TIME_NANOS to SLEEP_TIME_MILLIS, and changed from
parkNanos to Thread.sleep - high resolution is not required, and this
allows the watcher thread to be interrupted to shut it down.